requirements.txt
click==8.0.1
Flask==2.0.1
Flask-MySQLdb==0.2.0
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
pylint==2.8.3
PyMySQL==1.0.2
python-dateutil==2.8.1
rope==0.19.0
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
